+++ base/client.jl
@@ -361,7 +361,7 @@ function load_machine_file(path::AbstractString)
         s = split(line, '*'; keep = false)
         map!(strip, s, s)
         if length(s) > 1
-            cnt = isnumber(s[1]) ? parse(Int,s[1]) : Symbol(s[1])
+            cnt = all(isdigit, s[1]) ? parse(Int,s[1]) : Symbol(s[1])
             push!(machines,(s[2], cnt))
         else
             push!(machines,line)